refactor: move save/load logic to SaveLoadHandler class
